The .createSynthDefInstrument() method is now clever/presumptuous enough to detect if a SynthDef should be played monophonicly or polyphonicly and also deals with mono/stereo SynthDefs sensibly.
A single Synth will be created if the SynthDef passed in doesn't free itself
(has no doneAction) and it's parameters changed each time a note comes in.

If the SynthDef passed in does free itself it will be set up such that
multiple Synths are created and can be played polyphonicly.

Also did a little refactoring.